Separador implementado en Vaca_Toro y eliminacion de carpetas innecesarias por REFACTORIZACION
