--- src/Controlar.js
@@ -1,10 +1,0 @@
-import Buscar_Vaca_Toro from "./Vaca_Toro.js"
-
-function Juego_Vaca_Toro(cod1,cod2,cant) {
-
-    let resp;
-
-    for(var i=cant; i>0; i--) {
-        resp = Buscar_Vaca_Toro(cod1,cod2);        
-    }        
-}
