+++ src/Game.js
@@ -1,4 +1,7 @@
 class game{
 
-    
+    constructor() {
+
+        let intentos = 10;
+    }
 }
