--- src/Jugador_A.js
@@ -1,9 +1,0 @@
-function Validar_Codigo(codigo) {
-
-    if(codigo == "") {
-        return "Codigo Vacio";    
-    }
-    return codigo;
-}
-
-export default Validar_Codigo;
