--- src/Mensajes.js
@@ -1,11 +1,0 @@
-class Mensajes {
-
-    Mensaje_Felicidades() {
-        return "Felicidades Adivinaste El codigo";
-    }
-    Mensaje_Felicidades() {
-        return "Codigo Vacio";
-    }
-}
-
-export default Mensajes;
