+++ src/Vaca_Toro.spec.js
@@ -1,21 +1,22 @@
-import Validar_Codigo from "./Jugador_A.js"
-//import compararCodigo from "./vacaToro"
-import Separar_Codigo from "./Separador.js"
 import Vaca_Toro from "./Vaca_Toro.js"
 
 describe("CODIGO SECRETO - 'JUGADOR A'", () => {
 // Alfred Garcia
   it("Cuando jugador A ingrese un codigo secreto vacio, deberia devolver codigo vacio ", () => {
-    expect(Validar_Codigo("")).toEqual("Codigo Vacio");
+    let juego = new Vaca_Toro();
+    expect(juego.Validar_Codigo("")).toEqual("Codigo Vacio");
   });
   it("Cuando jugador A ingrese un codigo secreto de 4 digitos, deberia devolver dicho codigo ", () => {
-    expect(Validar_Codigo(1206)).toEqual(1206);
+    let juego = new Vaca_Toro
+    expect(juego.Validar_Codigo(1206)).toEqual(1206);
   });
   it("Cuando jugador A ingrese un codigo secreto de 5 digitos, deberia devolver dicho codigo ", () => {
-    expect(Validar_Codigo(45782)).toEqual(45782);
+    let juego = new Vaca_Toro
+    expect(juego.Validar_Codigo(45782)).toEqual(45782);
   });
   it("Cuando jugador A ingrese un codigo secreto de 6 digitos, deberia devolver dicho codigo ", () => {
-    expect(Validar_Codigo(123456)).toEqual(123456);
+    let juego = new Vaca_Toro
+    expect(juego.Validar_Codigo(123456)).toEqual(123456);
   });
   it("Si jugador A ingresa numero 3412 y el B ingresa 7604 deberia retornar * ", () => {    
     let juego = new Vaca_Toro();
@@ -66,8 +67,9 @@ describe("CODIGO SECRETO - 'JUGADOR A'", () => {
   
 describe("'JUGADOR B'", () => {
 // Marcelo Salinas
-  it("Cuando jugador B ingrese un codigo secreto vacio para adivinar el codigo, deberia mostrar, 'Ingresaste codigo vacio'  ", () => {    
-    expect(Validar_Codigo("")).toEqual("Codigo Vacio");
+  it("Cuando jugador B ingrese un codigo secreto vacio para adivinar el codigo, deberia mostrar, 'Ingresaste codigo vacio'  ", () => {  
+    let juego = new Vaca_Toro  
+    expect(juego.Validar_Codigo("")).toEqual("Codigo Vacio");
   });
 // Alfred Garcia
 });
@@ -75,10 +77,12 @@ describe("'JUGADOR B'", () => {
 describe("'AMBOS JUGADORES'", () => {
 // Alfred Garcia
   it("Si jugador ingresa el numero 5786 este debe devolver el numero en vector [5,7,8,6]", () => {    
-    expect(Separar_Codigo(5786)).toEqual(["5", "7", "8", "6"]);
+    let juego = new Vaca_Toro
+    expect(juego.Separar_Codigo(5786)).toEqual(["5", "7", "8", "6"]);
   });
   it("Si jugador ingresa numero 457896 este debe devolver el numero en vector [4,5,7,8,9,6]", () => {    
-    expect(Separar_Codigo(457896)).toEqual(["4", "5", "7", "8", "9", "6"]);
+    let juego = new Vaca_Toro
+    expect(juego.Separar_Codigo(457896)).toEqual(["4", "5", "7", "8", "9", "6"]);
   });
 // Marcelo Salinas
 });
